Addition to save and render trained networks
Use -S to save a network.
Use -R to render a saved network.

Need to figure out where it would be best to save the network to. But that should not be a lot of work.
Issue with this method is that changes to environments can not be made -> 7x7 to 10x10 will give issues. Speed can be changed.