#22 Reformulate policy loss to gradient
Works with wandb

What we previously thought of as "loss" is actually the policy gradient which we do in fact want to _ascend_.
The value loss is a separate metric and should be treated as such.

Just do a random run of the AC in a new project and see what we are tracking now.

Progressbar tracking is also updated (note that for AC runs, you will need to use a terminal window of around 90-100 columns, 80 is not enough for all three metrics.